Refactor Flask project into modular blueprints
